feat(formacion): rediseno con paleta unificada y CourseCard extraido
- Cards de Onboarding con gradientes rotativos (indigo/cian/purpura/teal)
- Card "Continuar aprendiendo" como banner gradiente colorido
- Boton Empezar/Certificado compacto alineado a la derecha
- Boton Editar como pill con gradiente segun el color de la card
- CourseCard extraido a components/formacion/CourseCard.tsx
- Helpers compartidos en lib/formacion-helpers.ts
- Fallback para URLs storage.example.com que no resuelven
- Anadido jspdf-autotable a dependencias
